feat(verify): diagnostics regression veto — new errors block fast-submit even when the flip holds (#861)
A flipped witness plus a fresh type error in an untested module could
still fast-submit: lint is rightly excluded from the flip oracle, but it
could not veto either. Now it can — and only that. A new LintProbe port
(host-side: the workspace's own diagnostics plan — cargo check / tsc /
eslint / ruff, closed vocabulary, fixed argv — via a new public
stella_tools::diagnostics::snapshot) feeds a set-difference against the
pre-execution baseline; new ERRORS drop the SubmitFast rung to the
judge, new warnings only when diagnostics_veto_warnings opts in.

Cost discipline: the delta runs inside the pre-submit audit, so only a
run about to claim deterministic credit pays it — lint before the #859
confirmation (cheaper first; a veto makes the confirmation moot).
Isolated candidates take their baseline lazily from the still-pristine
session tree (zero lint spend unless a fast-submit is imminent); the
in-place path snapshots eagerly because execution destroys its baseline.
Record identity excludes line/col so shifted pre-existing diagnostics
don't read as new; the judge sees at most a 3-line sample of new
errors, not the linter's whole opinion. Degrades open on every
unavailable path — the veto may only ever withhold, never invent.

Also splits verify.rs tests to verify/tests.rs and the probe helpers to
pipeline/verify_probes.rs (file-size ratchet), and moves the hardening
e2e scenarios to pipeline/tests/verification_hardening.rs.
