+++ scripts/file-size-baseline.txt
@@ -16,9 +16,9 @@
 4173 bench/terminal_bench_analysis/tests/test_tb21_analysis.py
 1615 bench/terminal_bench_analysis/tests/test_tb21_evidence_contract.py
 1789 stella-cli/src/agent_tests.rs
-2298 stella-cli/src/agent.rs
+2299 stella-cli/src/agent.rs
 1576 stella-cli/src/candidate_ws.rs
-4388 stella-cli/src/command_deck.rs
+4389 stella-cli/src/command_deck.rs
 2095 stella-core/src/bus.rs
 2195 stella-core/src/driver.rs
 2965 stella-core/src/driver/tests.rs
@@ -27,8 +27,8 @@
 1784 stella-model/src/bedrock.rs
 1952 stella-model/src/openai.rs
 1685 stella-model/src/zai/tests.rs
-2886 stella-pipeline/src/pipeline.rs
-2277 stella-pipeline/src/pipeline/tests.rs
+2989 stella-pipeline/src/pipeline.rs
+2332 stella-pipeline/src/pipeline/tests.rs
 2603 stella-protocol/src/event.rs
 2055 stella-store/src/lib.rs
 2204 stella-store/src/tests.rs
