+++ stella-cli/src/agent.rs
@@ -419,6 +419,7 @@ async fn run_pipeline_one_shot(
             touches: &crate::agent::RegistryTouches(&registry),
             diagnostics: &ws_ports.diagnostic_runner,
             tests: &ws_ports.test_runner,
+            lint: Some(&ws_ports.lint_probe),
             approvals: if approval_capability == PipelineApprovalCapability::Stdio {
                 &stdio_gate
             } else {
