+++ stella-cli/src/agent/goal.rs
@@ -681,6 +681,7 @@ async fn run_goal_pipeline_turn(
                 touches: &crate::agent::RegistryTouches(registry),
                 diagnostics: &ws_ports.diagnostic_runner,
                 tests: &ws_ports.test_runner,
+                lint: Some(&ws_ports.lint_probe),
                 approvals: &HEADLESS_APPROVAL_GATE,
                 sleeper: &TokioSleeper,
                 hooks: cfg
