+++ stella-cli/src/fleet_cmd.rs
@@ -802,6 +802,7 @@ async fn run_task(
                 touches: &crate::agent::RegistryTouches(&registry),
                 diagnostics: &ws_ports.diagnostic_runner,
                 tests: &ws_ports.test_runner,
+                lint: Some(&ws_ports.lint_probe),
                 approvals: &agent::HEADLESS_APPROVAL_GATE,
                 sleeper: &TokioSleeper,
                 hooks: cfg
