+++ stella-pipeline/src/lib.rs
@@ -112,9 +112,9 @@ pub use pipeline::{
 pub use ports::{
     AdoptedChange, AlwaysAbortGate, ApprovalGate, ArtifactIdentity, ArtifactKind,
     CandidateWorkspace, CandidateWorkspacePort, CmdKind, CmdOutcome, ContextRecallPort,
-    DiagnosticInvocation, DiagnosticRunner, FileTouchPort, McpPrefetchPort, NoContextRecall,
-    NoFileTouches, NoRepoStatus, NoRepoStructure, PipelinePorts, ProviderResolver, Recall,
-    RecalledFrame, RepoStatusPort, RepoStructurePort, ScopeDecision, StdioApprovalGate,
+    DiagnosticInvocation, DiagnosticRunner, FileTouchPort, LintProbe, LintRecord, McpPrefetchPort,
+    NoContextRecall, NoFileTouches, NoRepoStatus, NoRepoStructure, PipelinePorts, ProviderResolver,
+    Recall, RecalledFrame, RepoStatusPort, RepoStructurePort, ScopeDecision, StdioApprovalGate,
     TestInvocation, TestRunner, WorkspaceError,
 };
 pub use triage::TaskClass;
