+++ stella-pipeline/src/pipeline/tests/best_of_n.rs
@@ -119,6 +119,7 @@ async fn single_shot_never_touches_the_candidate_workspace_port() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -280,6 +281,7 @@ async fn best_of_n_without_a_port_degrades_to_the_shared_tree_with_a_warning() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -356,6 +358,7 @@ async fn a_mid_turn_steer_reaches_every_candidate_not_just_the_first() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -423,6 +426,7 @@ async fn the_steer_is_injected_once_per_candidate_not_once_per_step() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -492,6 +496,7 @@ async fn a_fan_out_narrates_each_candidate_and_names_the_winner() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -563,6 +568,7 @@ async fn a_single_candidate_run_emits_no_fan_out_narration() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -650,6 +656,7 @@ async fn every_isolated_candidate_also_sees_the_steer() {
             touches: &NoFileTouches,
             diagnostics: &diagnostics,
             tests: &diagnostics,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
