+++ stella-pipeline/src/pipeline/tests/golden.rs
@@ -106,6 +106,7 @@ async fn golden_single_task_deterministic_flip() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -171,6 +172,7 @@ async fn golden_judge_escalation_without_a_test_command() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
