+++ stella-pipeline/src/pipeline/tests/management_accounting.rs
@@ -69,6 +69,7 @@ async fn late_triage_is_abandoned_and_reported_incomplete() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -150,6 +151,7 @@ async fn triage_budget_crossing_aborts_before_a_second_provider_call() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
