+++ stella-pipeline/src/pipeline/tests/mcp_prefetch.rs
@@ -46,6 +46,7 @@ async fn best_of_n_folds_the_mcp_prefetch_into_every_candidate() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -109,6 +110,7 @@ async fn a_prefetch_miss_never_aborts_the_run() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
