+++ stella-pipeline/src/pipeline/tests/scope_gate_interactive.rs
@@ -77,6 +77,7 @@ async fn interactive_scope_review_approve_proceeds_past_the_gate() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -166,6 +167,7 @@ async fn a_revision_note_re_plans_and_gates_again() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -259,6 +261,7 @@ async fn endless_revision_notes_stop_at_the_cap_and_say_why() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
