+++ stella-pipeline/src/pipeline/tests/terminal_outcomes.rs
@@ -49,6 +49,7 @@ async fn red_final_verdict_is_verification_failed_not_completed() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -130,6 +131,7 @@ async fn enforced_budget_breach_in_triage_stops_before_the_next_paid_stage() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -201,6 +203,7 @@ async fn unavailable_independent_witness_degrades_instead_of_aborting() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
