+++ stella-pipeline/src/pipeline/tests/usage.rs
@@ -93,6 +93,7 @@ async fn run_triage_only(
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -206,6 +207,7 @@ async fn plan_and_plan_repair_each_emit_one_paid_call_envelope() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -307,6 +309,7 @@ async fn distress_guidance_and_engine_revisions_keep_distinct_envelopes() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -362,6 +365,7 @@ async fn model_judge_call_is_metered_separately_from_worker() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
