+++ stella-pipeline/src/pipeline/tests/warrant.rs
@@ -46,6 +46,7 @@ async fn a_docs_only_change_completes_without_buying_a_judge_call() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -145,6 +146,7 @@ async fn a_blind_diff_probe_never_completes_as_nothing_changed() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -247,6 +249,7 @@ async fn the_judge_is_told_what_the_recorder_counted_not_zero() {
             touches: &touches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -313,6 +316,7 @@ async fn a_non_repository_says_so_rather_than_reporting_a_generic_failure() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -326,6 +330,7 @@ async fn a_non_repository_says_so_rather_than_reporting_a_generic_failure() {
     let surface = CandidateSurface {
         diagnostics: &runner,
         tests: &runner,
+        lint: None,
         repo_status: &repo_status,
         cwd: None,
         hook_runner: None,
@@ -395,6 +400,7 @@ async fn a_turn_that_called_no_tool_does_not_report_success() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -493,6 +499,7 @@ async fn a_no_op_turn_is_sent_back_to_do_the_work() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
