+++ stella-pipeline/src/pipeline/tests/witness_isolation.rs
@@ -135,6 +135,7 @@ async fn witness_worker_and_revision_hooks_are_bound_to_the_candidate_root() {
             touches: &NoFileTouches,
             diagnostics: &session_runner,
             tests: &session_runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: Some((&hooks, &hook_runner)),
@@ -225,6 +226,7 @@ async fn authored_witness_degrades_when_judge_is_worker() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
@@ -467,6 +469,7 @@ async fn authored_witness_isolation_failure_degrades_to_a_bare_run() {
             touches: &NoFileTouches,
             diagnostics: &runner,
             tests: &runner,
+            lint: None,
             approvals: &approvals,
             sleeper: &sleeper,
             hooks: None,
