+++ stella-pipeline/src/verify.rs
@@ -274,7 +274,11 @@ pub enum LadderDecision {
 
 /// The evidence gathered after execution, over which [`ladder_decision`]
 /// reasons. All fields are owned plain data — the ladder is a pure function.
-#[derive(Debug, Clone, Copy, PartialEq)]
+///
+/// `Default` is the all-dark input: nothing observed, nothing dispatched,
+/// budget zero. Tests name the fields they exercise and take the rest from
+/// it, so adding an evidence channel does not rewrite every literal.
+#[derive(Debug, Clone, Copy, PartialEq, Default)]
 pub struct LadderInputs {
     /// The flip oracle reached `Flipped` for its tracked test command.
     pub flip_achieved: bool,
@@ -312,6 +316,22 @@ pub struct LadderInputs {
     /// nearly all real work lands — so an unrecognized name must never be the
     /// reason a turn is written off as a no-op.
     pub mutating_actions: u32,
+    /// Lint/typecheck ERRORS the candidate introduced over the baseline
+    /// snapshot (#861) — the regression veto's trigger. `0` both when the
+    /// tree is clean and when the probe is unavailable: lint can only ever
+    /// *withhold* a fast-submit (rung 4), never grant one, so an absent
+    /// probe degrades to the pre-veto ladder rather than fabricating a
+    /// regression.
+    pub new_diag_errors: u32,
+    /// New lint/typecheck warnings over the baseline. Vetoes only when
+    /// [`Self::veto_warnings`] opts in — a fresh warning is real signal but
+    /// blocking on it by default would tax every workspace with a chatty
+    /// linter.
+    pub new_diag_warnings: u32,
+    /// Configuration, not evidence: whether new warnings also veto the
+    /// fast-submit (errors always do). Carried here so the ladder stays a
+    /// pure function of one input value.
+    pub veto_warnings: bool,
 }
 
 impl LadderInputs {
@@ -395,10 +415,18 @@ pub fn ladder_decision(inputs: &LadderInputs) -> LadderDecision {
     if inputs.evidence_is_blind() {
         return LadderDecision::Unverifiable;
     }
-    // 4. Full deterministic pass — submit fast, judge skipped.
+    // 4. Full deterministic pass — submit fast, judge skipped. The
+    //    diagnostics conjuncts are the regression veto (#861): a flipped
+    //    witness plus a fresh type error in an untested module is exactly
+    //    the inconclusive case the judge exists for, so new errors (and,
+    //    opted-in, new warnings) drop this rung through to escalation. Lint
+    //    stays excluded from the oracle — it can veto a submit, never
+    //    verify one.
     if inputs.flip_achieved
         && inputs.touched_tests_passed == Some(true)
         && inputs.diff_lines <= inputs.diff_budget
+        && inputs.new_diag_errors == 0
+        && (!inputs.veto_warnings || inputs.new_diag_warnings == 0)
     {
         return LadderDecision::SubmitFast;
     }
@@ -606,801 +634,4 @@ pub fn guidance_prompt(goal: &str, diff: &str, evidence_summary: &str) -> String
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use proptest::prelude::*;
-
-    // FlipOracle transitions
-
-    #[test]
-    fn a_pass_with_no_prior_failure_proves_nothing() {
-        let mut oracle = FlipOracle::new();
-        assert_eq!(
-            oracle.observe("cargo test -p x", true),
-            ObserveOutcome::NoEvidence
-        );
-        assert_eq!(oracle.state(), FlipState::None);
-        assert!(!oracle.is_flipped());
-        // It didn't even lock the command.
-        assert_eq!(oracle.tracked_command(), None);
-    }
-
-    #[test]
-    fn fail_then_pass_of_the_same_command_flips() {
-        let mut oracle = FlipOracle::new();
-        assert_eq!(
-            oracle.observe("cargo test -p x", false),
-            ObserveOutcome::Advanced
-        );
-        assert_eq!(oracle.state(), FlipState::Failing);
-        assert_eq!(
-            oracle.observe("cargo test -p x", true),
-            ObserveOutcome::Advanced
-        );
-        assert!(oracle.is_flipped());
-    }
-
-    #[test]
-    fn whitespace_differences_are_the_same_tracked_command() {
-        let mut oracle = FlipOracle::new();
-        oracle.observe("cargo   test  -p   x", false);
-        // Re-run with normalized whitespace — must be recognized as the same.
-        let out = oracle.observe("cargo test -p x", true);
-        assert_eq!(out, ObserveOutcome::Advanced);
-        assert!(oracle.is_flipped());
-    }
-
-    #[test]
-    fn a_pass_of_a_different_command_never_flips() {
-        let mut oracle = FlipOracle::new();
-        oracle.observe("cargo test -p a", false);
-        // A pass of a DIFFERENT command must be ignored — proving a's failure
-        // fixed is not established by b passing.
-        assert_eq!(
-            oracle.observe("cargo test -p b", true),
-            ObserveOutcome::Ignored
-        );
-        assert!(!oracle.is_flipped());
-        assert_eq!(oracle.state(), FlipState::Failing);
-    }
-
-    #[test]
-    fn flipped_regresses_honestly_if_the_command_fails_again() {
-        let mut oracle = FlipOracle::new();
-        oracle.observe("t", false);
-        oracle.observe("t", true);
-        assert!(oracle.is_flipped());
-        // A later failure of the same command honestly moves back to Failing.
-        oracle.observe("t", false);
-        assert_eq!(oracle.state(), FlipState::Failing);
-        assert!(!oracle.is_flipped());
-    }
-
-    // ── Confirmation run (#859) ──────────────────────────────────────────
-
-    #[test]
-    fn confirmation_pass_keeps_the_flip() {
-        let mut oracle = FlipOracle::new();
-        oracle.observe("t", false);
-        oracle.observe("t", true);
-        assert!(oracle.is_flipped());
-        oracle.confirm(true);
-        assert!(oracle.is_flipped());
-        assert!(!oracle.is_unstable());
-    }
-
-    #[test]
-    fn confirmation_fail_makes_the_flip_unstable() {
-        // A flaky test: fail → pass (flip), but the confirmation re-run
-        // fails. The oracle must NOT credit a deterministic pass.
-        let mut oracle = FlipOracle::new();
-        oracle.observe("t", false);
-        oracle.observe("t", true);
-        assert!(oracle.is_flipped());
-        oracle.confirm(false);
-        assert!(!oracle.is_flipped(), "an unconfirmed flip is not Flipped");
-        assert!(oracle.is_unstable());
-    }
-
-    #[test]
-    fn confirm_outside_flipped_is_a_noop() {
-        // Confirmation is only meaningful where a flip stands to be credited.
-        let mut oracle = FlipOracle::new();
-        oracle.confirm(false);
-        assert_eq!(oracle.state(), FlipState::None);
-        oracle.observe("t", false); // Failing
-        oracle.confirm(false);
-        assert_eq!(oracle.state(), FlipState::Failing);
-        assert!(!oracle.is_unstable());
-    }
-
-    #[test]
-    fn a_flaky_test_that_passes_again_recovers_from_unstable() {
-        // A later pass of the tracked command re-flips; the pipeline's
-        // pre-submit audit will demand a FRESH confirmation before crediting
-        // that new flip, so recovery never skips the guard.
-        let mut oracle = FlipOracle::new();
-        oracle.observe("t", false);
-        oracle.observe("t", true);
-        oracle.confirm(false);
-        assert!(oracle.is_unstable());
-        oracle.observe("t", true);
-        assert!(oracle.is_flipped(), "a pass after Unstable re-flips");
-        assert!(!oracle.is_unstable());
-    }
-
-    #[test]
-    fn unstable_stays_unstable_on_another_failure() {
-        let mut oracle = FlipOracle::new();
-        oracle.observe("t", false);
-        oracle.observe("t", true);
-        oracle.confirm(false);
-        oracle.observe("t", false);
-        assert_eq!(oracle.state(), FlipState::Unstable);
-        assert!(!oracle.is_flipped());
-    }
-
-    #[test]
-    fn unstable_ignores_other_commands_like_every_locked_state() {
-        let mut oracle = FlipOracle::new();
-        oracle.observe("a", false);
-        oracle.observe("a", true);
-        oracle.confirm(false);
-        assert_eq!(oracle.observe("b", true), ObserveOutcome::Ignored);
-        assert_eq!(oracle.state(), FlipState::Unstable);
-    }
-
-    /// The binding #859 property at the ladder level: an unconfirmed flip
-    /// must escalate, never fast-submit.
-    #[test]
-    fn an_unstable_oracle_does_not_credit_a_deterministic_pass() {
-        let mut oracle = FlipOracle::new();
-        oracle.observe("t", false);
-        oracle.observe("t", true);
-        oracle.confirm(false);
-        let inputs = LadderInputs {
-            flip_achieved: oracle.is_flipped(), // false
-            touched_tests_passed: Some(true),
-            diff_lines: 5,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 1,
-            mutating_actions: 1,
-        };
-        assert_eq!(
-            ladder_decision(&inputs),
-            LadderDecision::ModelJudge,
-            "an unconfirmed flip must escalate to the judge, not SubmitFast"
-        );
-    }
-
-    #[test]
-    fn from_none_a_single_observation_can_never_reach_flipped() {
-        // The core invariant, in the small: you cannot go None → Flipped in
-        // one step; Failing is mandatory.
-        for passed in [true, false] {
-            let mut oracle = FlipOracle::new();
-            oracle.observe("t", passed);
-            assert_ne!(
-                oracle.state(),
-                FlipState::Flipped,
-                "one observation from None must never be Flipped"
-            );
-        }
-    }
-
-    // normalize_command
-
-    #[test]
-    fn normalize_collapses_whitespace_and_trims_but_keeps_order() {
-        assert_eq!(normalize_command("  a   b\tc \n"), "a b c");
-        // Order is preserved (not sorted) — reordering could change meaning.
-        assert_ne!(normalize_command("a b"), normalize_command("b a"));
-    }
-
-    // ladder_decision
-
-    #[test]
-    fn red_touched_tests_revise_without_a_judge() {
-        let decision = ladder_decision(&LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: Some(false),
-            diff_lines: 3,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 0,
-            mutating_actions: 1,
-        });
-        assert_eq!(decision, LadderDecision::Revise);
-    }
-
-    #[test]
-    fn full_deterministic_pass_submits_fast() {
-        let decision = ladder_decision(&LadderInputs {
-            flip_achieved: true,
-            touched_tests_passed: Some(true),
-            diff_lines: 40,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 0,
-            mutating_actions: 1,
-        });
-        assert_eq!(decision, LadderDecision::SubmitFast);
-    }
-
-    /// The Terminal-Bench trial from #973, reconstructed as ladder inputs:
-    /// the task directory is not a git repository so `git diff` cannot read it,
-    /// there is no test command so the oracle never armed and no tests ran, and
-    /// (before the fix) the recorder's touches never reached the pipeline. All
-    /// four channels dark at once.
-    #[test]
-    fn every_channel_blind_abstains_instead_of_asking_a_judge_to_guess() {
-        let inputs = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 0,
-            diff_budget: 100,
-            diff_available: false,
-            file_change_events: 0,
-            mutating_actions: 1,
-        };
-        assert!(inputs.evidence_is_blind());
-        assert_eq!(
-            ladder_decision(&inputs),
-            LadderDecision::Unverifiable,
-            "an unobservable turn must abstain, never buy a judge call to guess"
-        );
-        let evidence = unverifiable_evidence(&inputs);
-        assert!(evidence.summary.contains("UNVERIFIABLE"));
-        assert!(
-            !evidence.deterministic,
-            "an absent result must never wear the deterministic badge"
-        );
-    }
-
-    /// The single signal that rescues the blind case, and the one the bug
-    /// suppressed: the recorder saw the tree change even though nothing could
-    /// render *how*. That is real evidence, so the ladder must escalate rather
-    /// than abstain — and the judge is told a positive fact instead of a zero.
-    #[test]
-    fn a_recorded_file_touch_is_evidence_even_with_no_readable_diff() {
-        let inputs = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 0,
-            diff_budget: 100,
-            diff_available: false,
-            file_change_events: 6,
-            mutating_actions: 1,
-        };
-        assert!(
-            !inputs.evidence_is_blind(),
-            "six observed mutations are not an absence of evidence"
-        );
-        assert_eq!(ladder_decision(&inputs), LadderDecision::ModelJudge);
-    }
-
-    /// A red test is a real observation, so a turn carrying one is never blind
-    /// however dark everything else is — and `Revise` still wins, because a
-    /// deterministic failure must not be softened into an abstention.
-    #[test]
-    fn a_red_test_outranks_blindness() {
-        let inputs = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: Some(false),
-            diff_lines: 0,
-            diff_budget: 100,
-            diff_available: false,
-            file_change_events: 0,
-            mutating_actions: 1,
-        };
-        assert!(!inputs.evidence_is_blind());
-        assert_eq!(ladder_decision(&inputs), LadderDecision::Revise);
-    }
-
-    /// The distinction the whole rung exists for: "I looked and saw nothing"
-    /// (an available probe reporting a zero-line diff) is inconclusive and
-    /// worth a judge; "I could not look" is not.
-    #[test]
-    fn a_readable_empty_diff_is_not_blindness() {
-        let inputs = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 0,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 0,
-            mutating_actions: 1,
-        };
-        assert!(!inputs.evidence_is_blind());
-        assert_eq!(ladder_decision(&inputs), LadderDecision::ModelJudge);
-    }
-
-    /// The `regex-log` Terminal-Bench 2.1 trial, reconstructed as ladder
-    /// inputs: `glm-5.2` emitted 123 reasoning events, called no tool at all,
-    /// and the run reported success on a task it never touched (Harbor scored
-    /// it 0.0). Every input below is read off that trial's own event stream.
-    ///
-    /// The assertion that matters is the second one. This turn *is* blind by
-    /// [`LadderInputs::evidence_is_blind`] — all four channels are dark, and
-    /// they were dark before the fix too — so the bug was never a wrong
-    /// blindness check. It was that blindness was asked first, and abstaining
-    /// answered `passed: true`. Ordering is the entire fix, which is why this
-    /// test pins both facts at once.
-    #[test]
-    fn a_turn_that_called_no_tool_is_a_no_op_not_an_abstention() {
-        let inputs = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 0,
-            diff_budget: 400,
-            // Not a git repository, so the probe is permanently dark here.
-            diff_available: false,
-            file_change_events: 0,
-            mutating_actions: 0,
-        };
-        assert!(inputs.nothing_was_attempted());
-        assert!(
-            inputs.evidence_is_blind(),
-            "the no-op state satisfies the blind check too — the rung order is what separates them"
-        );
-        assert_eq!(
-            ladder_decision(&inputs),
-            LadderDecision::NothingAttempted,
-            "a turn that dispatched nothing must never reach the abstain rung and be passed"
-        );
-    }
-
-    /// The evidence a no-op turn carries, and the one word it must not borrow.
-    /// `UNVERIFIABLE` is a claim about the *instruments*; this turn's
-    /// instruments were fine and reported a determinate nothing, so it is
-    /// marked `deterministic: true` — the opposite of
-    /// [`unverifiable_evidence`], which is `false` precisely because it has no
-    /// result to report.
-    #[test]
-    fn a_no_op_verdict_is_a_result_not_a_missing_one() {
-        let inputs = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 0,
-            diff_budget: 400,
-            diff_available: false,
-            file_change_events: 0,
-            mutating_actions: 0,
-        };
-        let evidence = nothing_attempted_evidence(&inputs);
-        assert!(evidence.summary.contains("NO WORK ATTEMPTED"));
-        assert!(
-            !evidence.summary.contains("UNVERIFIABLE"),
-            "a determinate no-op must not describe itself as unobservable"
-        );
-        assert!(
-            evidence.deterministic,
-            "the pipeline counted its own dispatches; no probe could have said otherwise"
-        );
-    }
-
-    /// The case the abstain rung exists for, and the reason the fix above is a
-    /// new rung rather than a flipped boolean: `log-summary-date-ranges` ran
-    /// eight tool calls, wrote its answer through shell redirects (so the
-    /// recorder logged no touch), could not be diffed — and scored **1.0**
-    /// against its Harbor verifier. Identical dark channels, opposite truth.
-    /// Only `mutating_actions` tells it apart from the trial above.
-    #[test]
-    fn a_blind_turn_that_did_act_still_abstains() {
-        let inputs = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 0,
-            diff_budget: 400,
-            diff_available: false,
-            file_change_events: 0,
-            mutating_actions: 8,
-        };
-        assert!(inputs.evidence_is_blind());
-        assert!(
-            !inputs.nothing_was_attempted(),
-            "eight dispatched mutating calls are not an absence of work"
-        );
-        assert_eq!(
-            ladder_decision(&inputs),
-            LadderDecision::Unverifiable,
-            "work nothing could observe must still abstain — failing it closed would report a \
-             correct run as broken, and no revision could ever clear it"
-        );
-    }
-
-    /// Any single positive observation outranks the no-op rung, whatever the
-    /// dispatch count says. Something changed; something changed is a thing to
-    /// explain, and this shortcut does not get to skip explaining it.
-    #[test]
-    fn one_positive_observation_defeats_the_no_op_rung() {
-        let base = LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 0,
-            diff_budget: 400,
-            diff_available: false,
-            file_change_events: 0,
-            mutating_actions: 0,
-        };
-        assert!(base.nothing_was_attempted());
-
-        let recorded_touch = LadderInputs {
-            file_change_events: 1,
-            ..base
-        };
-        let visible_diff = LadderInputs {
-            diff_lines: 12,
-            ..base
-        };
-        let flipped = LadderInputs {
-            flip_achieved: true,
-            ..base
-        };
-        let tests_ran = LadderInputs {
-            touched_tests_passed: Some(true),
-            ..base
-        };
-        for (label, inputs) in [
-            ("a recorded file touch", recorded_touch),
-            ("a non-empty diff", visible_diff),
-            ("a flip", flipped),
-            ("a test result", tests_ran),
-        ] {
-            assert!(
-                !inputs.nothing_was_attempted(),
-                "{label} is a positive observation; the turn is not a no-op"
-            );
-            assert_ne!(
-                ladder_decision(&inputs),
-                LadderDecision::NothingAttempted,
-                "{label} must route somewhere that reasons about it"
-            );
-        }
-    }
-
-    /// The judge is only ever asked when something could see — but when it is
-    /// asked, it must be told that a dark channel is a gap, not a finding. The
-    /// wild failure was a judge reading "the probe could not read the working
-    /// tree" as "the file does not exist".
-    #[test]
-    fn the_judge_prompt_forbids_reading_a_blind_probe_as_an_absence() {
-        let p = judge_prompt(
-            "fix the bug",
-            DIFF_PROBE_BLIND_SAMPLE,
-            "file_change_events=6",
-        );
-        assert!(
-            p.contains("could not read the working tree"),
-            "the prompt must name the failure mode it is guarding against"
-        );
-        assert!(p.contains("never on evidence you could not see"));
-    }
-
-    const DIFF_PROBE_BLIND_SAMPLE: &str =
-        "[the diff probe failed; the working tree could not be read.]";
-
-    #[test]
-    fn flip_and_green_but_over_diff_budget_escalates_to_judge() {
-        let decision = ladder_decision(&LadderInputs {
-            flip_achieved: true,
-            touched_tests_passed: Some(true),
-            diff_lines: 500,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 0,
-            mutating_actions: 1,
-        });
-        assert_eq!(
-            decision,
-            LadderDecision::ModelJudge,
-            "a large diff deserves a second opinion even with green tests"
-        );
-    }
-
-    #[test]
-    fn no_flip_evidence_escalates_to_judge_not_a_false_pass() {
-        // Tests green but never flipped (they always passed) → inconclusive.
-        let decision = ladder_decision(&LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: Some(true),
-            diff_lines: 5,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 0,
-            mutating_actions: 1,
-        });
-        assert_eq!(decision, LadderDecision::ModelJudge);
-    }
-
-    #[test]
-    fn tests_indeterminate_escalates_to_judge() {
-        let decision = ladder_decision(&LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: None,
-            diff_lines: 5,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 0,
-            mutating_actions: 1,
-        });
-        assert_eq!(decision, LadderDecision::ModelJudge);
-    }
-
-    // judge parsing + fallback
-
-    #[test]
-    fn parses_pass_and_fail_verdicts() {
-        assert_eq!(
-            parse_judge_response("PASS — looks correct").map(|v| v.passed),
-            Some(true)
-        );
-        assert_eq!(
-            parse_judge_response("FAIL: missing edge case").map(|v| v.passed),
-            Some(false)
-        );
-        assert_eq!(
-            parse_judge_response("Verdict: approved").map(|v| v.passed),
-            Some(true)
-        );
-        // A PASS line whose reasoning contains "no" must not be flipped to
-        // FAIL by an over-eager "no" match.
-        assert_eq!(
-            parse_judge_response("PASS — no obvious issues").map(|v| v.passed),
-            Some(true)
-        );
-        // Only the first non-empty line is authoritative.
-        assert_eq!(
-            parse_judge_response("FAIL\nthe change looks fine otherwise").map(|v| v.passed),
-            Some(false)
-        );
-    }
-
-    #[test]
-    fn unparseable_judge_response_is_none() {
-        assert_eq!(parse_judge_response("hmm, hard to say"), None);
-        assert_eq!(parse_judge_response(""), None);
-    }
-
-    #[test]
-    fn heuristic_fallback_passes_only_on_confirmed_green_tests() {
-        let green = heuristic_fallback(&LadderInputs {
-            flip_achieved: false,
-            touched_tests_passed: Some(true),
-            diff_lines: 0,
-            diff_budget: 100,
-            diff_available: true,
-            file_change_events: 0,
-            mutating_actions: 1,
-        });
-        assert!(green.passed);
-
-        for tests in [Some(false), None] {
-            let v = heuristic_fallback(&LadderInputs {
-                flip_achieved: true, // even a flip doesn't rescue an unconfirmed suite
-                touched_tests_passed: tests,
-                diff_lines: 0,
-                diff_budget: 100,
-                diff_available: true,
-                file_change_events: 0,
-                mutating_actions: 1,
-            });
-            assert!(!v.passed, "unconfirmed tests must fall back to FAIL");
-        }
-    }
-
-    #[test]
-    fn evidence_builders_tag_determinism_correctly() {
-        assert!(deterministic_pass_evidence(Some("cargo test"), 10).deterministic);
-        assert!(deterministic_fail_evidence("boom").deterministic);
-        let model = model_verdict_evidence(&JudgeVerdict {
-            passed: true,
-            reasoning: "looks fine".into(),
-        });
-        assert!(
-            !model.deterministic,
-            "model verdicts are never deterministic"
-        );
-    }
-
-    #[test]
-    fn judge_prompt_carries_goal_diff_and_evidence_but_asks_for_pass_fail() {
-        let p = judge_prompt("fix the bug", "@@ -1 +1 @@\n-x\n+y", "no flip; tests green");
-        assert!(p.contains("fix the bug"));
-        assert!(p.contains("+y"));
-        assert!(p.contains("no flip; tests green"));
-        assert!(p.contains("PASS"));
-        assert!(p.contains("FAIL"));
-    }
-
-    // The binding property (L-E11)
-
-    /// A reference oracle: replay a sequence of observations and independently
-    /// compute whether a genuine flip occurred (a failure of some command
-    /// followed later by a pass of that *same normalized* command, with no
-    /// intervening failure of it right before the pass). This mirrors the
-    /// state machine's intent so we can cross-check `FlipOracle` against it.
-    fn reference_flipped(observations: &[(String, bool)]) -> bool {
-        // Track the first command that fails (the oracle locks onto it).
-        let mut tracked: Option<String> = None;
-        let mut state = FlipState::None;
-        for (cmd, passed) in observations {
-            let norm = normalize_command(cmd);
-            match &tracked {
-                None => {
-                    if !passed {
-                        tracked = Some(norm);
-                        state = FlipState::Failing;
-                    }
-                }
-                Some(t) if *t == norm => {
-                    state = if *passed {
-                        FlipState::Flipped
-                    } else {
-                        FlipState::Failing
-                    };
-                }
-                _ => {}
-            }
-        }
-        matches!(state, FlipState::Flipped)
-    }
-
-    proptest! {
-        /// The binding invariant (L-E11): the oracle reports `Flipped` **only**
-        /// when the observation sequence contains a failing observation of the
-        /// tracked command strictly before the pass that flipped it. We prove
-        /// it two ways at once: (a) the live oracle agrees with an independent
-        /// reference computation, and (b) whenever the oracle is flipped, the
-        /// tracked command was observed failing at least once earlier in the
-        /// sequence.
-        #[test]
-        fn flip_requires_a_prior_failing_observation(
-            // A small alphabet of commands so collisions (same command re-run)
-            // actually happen; random pass/fail outcomes.
-            seq in prop::collection::vec(
-                ((0u8..4).prop_map(|n| format!("cargo test -p crate{n}")), any::<bool>()),
-                0..40,
-            )
-        ) {
-            let mut oracle = FlipOracle::new();
-            for (cmd, passed) in &seq {
-                oracle.observe(cmd, *passed);
-            }
-
-            // (a) Agreement with the independent reference.
-            prop_assert_eq!(oracle.is_flipped(), reference_flipped(&seq));
-
-            // (b) If flipped, the tracked command was seen failing earlier,
-            //     and a pass of it came after that failure.
-            if oracle.is_flipped() {
-                let tracked = oracle.tracked_command().expect("flipped implies a tracked command");
-                let norm_tracked = normalize_command(tracked);
-                let mut saw_fail = false;
-                let mut fail_before_pass = false;
-                for (cmd, passed) in &seq {
-                    if normalize_command(cmd) != norm_tracked {
-                        continue;
-                    }
-                    if !passed {
-                        saw_fail = true;
-                    } else if saw_fail {
-                        fail_before_pass = true;
-                    }
-                }
-                prop_assert!(saw_fail, "flipped without ever observing the tracked command fail");
-                prop_assert!(
-                    fail_before_pass,
-                    "flipped without a fail strictly before the flipping pass"
-                );
-            }
-        }
-
-        /// The oracle can never jump straight from `None` to `Flipped`: the
-        /// state after processing a prefix is `Flipped` only if the prefix
-        /// already contained a failure of the tracked command.
-        #[test]
-        fn never_none_to_flipped_in_one_step(
-            passed in any::<bool>(),
-        ) {
-            let mut oracle = FlipOracle::new();
-            oracle.observe("cargo test", passed);
-            prop_assert_ne!(oracle.state(), FlipState::Flipped);
-        }
-
-        /// #859: `confirm` can only ever demote. Under any interleaving of
-        /// observations and confirmations, `confirm(true)` never changes the
-        /// state and `confirm(false)` never leaves a flip standing — so the
-        /// confirmation guard cannot be gamed into MINTING deterministic
-        /// credit, only into withholding it.
-        #[test]
-        fn confirm_only_demotes(
-            seq in prop::collection::vec(
-                (
-                    (0u8..3).prop_map(|n| format!("cmd{n}")),
-                    any::<bool>(),
-                    prop::option::of(any::<bool>()),
-                ),
-                0..40,
-            )
-        ) {
-            let mut oracle = FlipOracle::new();
-            for (cmd, passed, confirmation) in &seq {
-                oracle.observe(cmd, *passed);
-                if let Some(confirmed) = confirmation {
-                    let state_before = oracle.state();
-                    oracle.confirm(*confirmed);
-                    if *confirmed {
-                        prop_assert_eq!(oracle.state(), state_before, "confirm(true) never moves the oracle");
-                    } else {
-                        prop_assert!(!oracle.is_flipped(), "confirm(false) never leaves a flip standing");
-                    }
-                }
-            }
-        }
-
-        /// The binding invariant of the no-op rung: **no ladder input that
-        /// dispatched nothing and observed nothing can resolve to a rung that
-        /// reports a pass.**
-        ///
-        /// Stated over the whole input space rather than the eleven trials,
-        /// because the defect was never about a particular trial's numbers —
-        /// it was that one reachable combination fell into `Unverifiable`,
-        /// which answers `passed: true`. A ranged sweep is what proves no such
-        /// combination is left, including ones nobody has produced yet.
-        ///
-        /// `diff_available` is swept and deliberately unconstrained: whether
-        /// the probe could look must make no difference to a turn that never
-        /// gave it anything to look at.
-        #[test]
-        fn a_turn_that_dispatched_nothing_never_reaches_a_passing_rung(
-            diff_available in any::<bool>(),
-            diff_budget in 0u32..1000,
-        ) {
-            let inputs = LadderInputs {
-                flip_achieved: false,
-                touched_tests_passed: None,
-                diff_lines: 0,
-                diff_budget,
-                diff_available,
-                file_change_events: 0,
-                mutating_actions: 0,
-            };
-            prop_assert!(inputs.nothing_was_attempted());
-            let decision = ladder_decision(&inputs);
-            prop_assert_eq!(decision, LadderDecision::NothingAttempted);
-            // The two rungs that answer `passed: true` in the pipeline. Named
-            // individually so a future rung that also passes has to be
-            // considered here rather than silently inheriting a green test.
-            prop_assert_ne!(decision, LadderDecision::SubmitFast);
-            prop_assert_ne!(decision, LadderDecision::Unverifiable);
-        }
-
-        /// The converse, and the guard on the fix's blast radius: any turn
-        /// that *did* dispatch a mutating call keeps the decision it had
-        /// before this rung existed. The no-op rung may only ever claim the
-        /// zero-dispatch corner.
-        #[test]
-        fn dispatching_anything_leaves_every_other_rung_untouched(
-            mutating_actions in 1u32..50,
-            flip_achieved in any::<bool>(),
-            touched in prop::option::of(any::<bool>()),
-            diff_lines in 0u32..800,
-            diff_available in any::<bool>(),
-            file_change_events in 0u32..50,
-        ) {
-            let inputs = LadderInputs {
-                flip_achieved,
-                touched_tests_passed: touched,
-                diff_lines,
-                diff_budget: 400,
-                diff_available,
-                file_change_events,
-                mutating_actions,
-            };
-            prop_assert!(!inputs.nothing_was_attempted());
-            prop_assert_ne!(ladder_decision(&inputs), LadderDecision::NothingAttempted);
-        }
-    }
-}
+mod tests;
